--- core/src/com/marama/game/Renderable.java
@@ -1,9 +1,0 @@
-package com.marama.game;
-
-public interface Renderable {
-    void resize(int width, int height);
-    void render();
-    void pause();
-    void resume();
-    void dispose();
-}
