+++ core/src/com/marama/view/screens/SplashScreen.java
@@ -1,0 +1,40 @@
+package com.marama.view.screens;
+
+import com.badlogic.gdx.Screen;
+
+public class SplashScreen implements Screen {
+    @Override
+    public void show() {
+
+    }
+
+    @Override
+    public void render(float delta) {
+
+    }
+
+    @Override
+    public void resize(int width, int height) {
+
+    }
+
+    @Override
+    public void pause() {
+
+    }
+
+    @Override
+    public void resume() {
+
+    }
+
+    @Override
+    public void hide() {
+
+    }
+
+    @Override
+    public void dispose() {
+
+    }
+}
