--- desktop/src/com/marama/game/desktop/DesktopLauncher.java
@@ -1,13 +1,0 @@
-package com.marama.game.desktop;
-
-import com.badlogic.gdx.backends.lwjgl.LwjglApplication;
-import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
-import com.marama.game.Game;
-
-
-public class DesktopLauncher {
-	public static void main (String[] arg) {
-		LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
-		new LwjglApplication(new Game(), config);
-	}
-}
