+++ desktop/src/com/marama/view/desktop/DesktopLauncher.java
@@ -1,0 +1,13 @@
+package com.marama.view.desktop;
+
+import com.badlogic.gdx.backends.lwjgl.LwjglApplication;
+import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
+import com.marama.view.View;
+
+
+public class DesktopLauncher {
+	public static void main (String[] arg) {
+		LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
+		new LwjglApplication(new View(), config);
+	}
+}
