+++ ios/robovm.properties
@@ -1,6 +1,6 @@
 app.version=1.0
-app.id=com.marama.game
-app.mainclass=com.marama.game.IOSLauncher
+app.id=com.marama.view
+app.mainclass=com.marama.view.IOSLauncher
 app.executable=IOSLauncher
 app.build=1
 app.name=marama-view
